Aplikacja_Webowa_Etap_5/client/requirements.txt
flask
uuid
pyjwt
redis
requests
six
authlib
requests
python-dotenv

